[!!!][TASK] Drop support for typo3 v11 and php 8.1
