+++ composer.json
@@ -27,10 +27,10 @@
 	],
 	"license": "GPL-2.0-or-later",
 	"require": {
-		"php": "^8.1",
+		"php": "^8.2",
 		"ext-json": "*",
-		"typo3/cms-core": "^11.5 || ^12.4 || ^13.4",
-		"typo3/cms-frontend": "^11.5 || ^12.4 || ^13.4",
+		"typo3/cms-core": "^12.4 || ^13.4",
+		"typo3/cms-frontend": "^12.4 || ^13.4",
 		"league/oauth2-client": "^2.8",
 		"firebase/php-jwt": "^6.10"
 	},
@@ -38,7 +38,7 @@
 		"phpunit/phpunit": "^10",
 		"typo3/testing-framework": "8.x-dev",
 		"typo3/minimal": "^13",
-		"typo3/cms-felogin": "^11.5 || ^12.4 || ^13.4"
+		"typo3/cms-felogin": "^12.4 || ^13.4"
 	},
 	"autoload": {
 		"psr-4": {
